+++ server/src/Producer.h
@@ -5,5 +5,5 @@ class Producer
 private:
     virtual void CreateConnection() = 0;
 public:
-    virtual void Report(int value) = 0;
+    virtual void Report(float value) = 0;
 };
